rag_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
chromadb==0.5.0
openai==1.30.1
pydantic==2.6.4
python-dotenv==1.0.1
tqdm==4.66.2
numpy==1.26.4
httpx==0.27.0